Update tests to use new nengo pytest setup
Switch tests to pytest-rng/allclose fixtures

Use simloader functionality for Simulator fixture

Switch to pytest_nengo for nengo test options

Use nodeid for nengo_test_unsupported
