+++ setup.cfg
@@ -46,8 +46,7 @@ max-complexity = 10
 max-line-length = 88
 
 [tool:pytest]
-addopts = -p nengo.tests.options --simulator nengo_dl.Simulator --ref-simulator nengo_dl.Simulator --disable-warnings
-xfail_strict = True
+addopts = --disable-warnings
 norecursedirs =
     .*
     *.egg
@@ -56,45 +55,45 @@ norecursedirs =
     docs
 filterwarnings =
     always
+xfail_strict = True
+nengo_simloader = nengo_dl.tests.make_test_sim
 nengo_test_unsupported =
-    nengo/tests/test_simulator.py:test_warn_on_opensim_del
+    tests/test_simulator.py::test_warn_on_opensim_del
         "nengo_dl raises a different (more visible) warning, see
-        tests/test_nengo_tests.py:test_warn_on_opensim_del"
-    nengo/tests/test_simulator.py:test_signal_init_values
+        tests/test_nengo_tests.py::test_warn_on_opensim_del"
+    tests/test_simulator.py::test_signal_init_values
         "different method required to manually step simulator, see
-        tests/test_nengo_tests.py:test_signal_init_values"
-    nengo/tests/test_simulator.py:test_entry_point
+        tests/test_nengo_tests.py::test_signal_init_values"
+    tests/test_simulator.py::test_entry_point
         "overridden so we can pass custom test simulators, see
-        tests/test_nengo_tests.py:test_entry_point"
-    nengo/tests/test_simulator.py:test_simulator_progress_bars
+        tests/test_nengo_tests.py::test_entry_point"
+    tests/test_simulator.py::test_simulator_progress_bars
         "nengo_dl uses a different progress bar system, see
-        tests/test_utils.py:test_progress_bar"
-    nengo/tests/test_simulator.py:test_dtype[*
+        tests/test_utils.py::test_progress_bar"
+    tests/test_simulator.py::test_dtype[*
         "nengo_dl uses a different system for signals/dtype, see
-        tests/test_nengo_tests.py:test_dtype"
-    nengo/tests/test_node.py:test_args
+        tests/test_nengo_tests.py::test_dtype"
+    tests/test_node.py::test_args
         "time is passed as np.float32, not a float, see
-        tests/test_nengo_tests.py:test_args"
-    nengo/tests/test_node.py:test_unconnected_node
+        tests/test_nengo_tests.py::test_args"
+    tests/test_node.py::test_unconnected_node
         "need to set `unroll_simulation` to ensure node runs the correct number of
-        times, see tests/test_nengo_tests.py:test_unconnected_node"
-    nengo/tests/test_synapses.py:test_alpha
-        "need to set looser tolerances due to float32 implementation, see
-        tests/test_processes.py:test_alpha"
-    nengo/tests/test_synapses.py:test_linearfilter
-        "need to set looser tolerances due to float32 implementation, see
-        tests/test_processes.py:test_linearfilter"
-    nengo/tests/test_ensemble.py:test_gain_bias
-        "use allclose instead of array_equal, see tests/test_simulator.py:test_gain_bias"
-    nengo/tests/test_transforms.py:test_convolution*
-        "need to set looser tolerances due to float32 implementation, see
-        tests/test_transforms.py:test_convolution"
-    nengo/tests/test_transforms.py:test_sparse[False-*
+        times, see tests/test_nengo_tests.py::test_unconnected_node"
+    tests/test_ensemble.py::test_gain_bias
+        "use allclose instead of array_equal, see
+        tests/test_nengo_tests.py::test_gain_bias"
+    tests/test_transforms.py::test_sparse[False-*
         "nengo-dl doesn't raise a warning for scipy=False, see
-        tests/test_nengo_tests.py:test_sparse"
-    nengo/tests/test_copy.py:test_pickle_sim[*
+        tests/test_nengo_tests.py::test_sparse"
+    tests/test_copy.py::test_pickle_sim[*
         "nengo-dl does not support pickling a Simulator, see
-        tests/test_simulator.py:test_pickle_error"
+        tests/test_simulator.py::test_pickle_error"
+allclose_tolerances =
+    tests/test_synapses.py::test_lowpass atol=5e-7
+    tests/test_synapses.py::test_triangle atol=5e-7
+    tests/test_synapses.py::test_alpha atol=5e-5
+    tests/test_synapses.py::test_linearfilter atol=1e-4
+    tests/test_transforms.py::test_convolution[* atol=1e-6
 
 [pylint]
 # note: pylint doesn't look in setup.cfg by default, need to call it with
